chore: move useBadgeOptions into a hook file
