+++ src/modules/settings/components/SettingKeywords.jsx
@@ -19,16 +19,15 @@ import {
 } from '@mantine/core';
 import {useDisclosure, useFocusTrap} from '@mantine/hooks';
 import classNames from 'classnames';
-import React, {useCallback, useEffect, useMemo, useRef, useState} from 'react';
+import React, {useCallback, useMemo, useRef, useState} from 'react';
 import Icon from '@/common/components/Icon';
+import useBadgeOptions from '@/common/hooks/BadgeOptions';
 import useCurrentChannel from '@/common/hooks/CurrentChannel';
 import usePortalRef from '@/common/hooks/PortalRef';
 import tableStyles from '@/common/styles/SettingEntryTable.module.css';
 import {openModal} from '@/common/utils/modal';
 import formatMessage from '@/i18n/index';
-import debug from '@/utils/debug';
 import {KeywordTypes} from '@/utils/keywords';
-import twitch from '@/utils/twitch';
 import ColorPicker from './ColorPicker';
 import Panel from './Panel';
 import styles from './SettingKeywords.module.css';
@@ -94,76 +93,6 @@ function openRegexGuideModal() {
     title: formatMessage({defaultMessage: 'Advanced Keywords'}),
     children: <RegexGuideModalBody />,
   });
-}
-
-// escapes only true regex metacharacters, unlike escapeRegExp, so the
-// keyword stays substring-matchable by mantine's built-in option filter
-function badgeKeyword(title) {
-  return `~/^${title.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')}$/`;
-}
-
-function buildBadgeOptions(globalBadges) {
-  const seenTitles = new Set();
-  const badgeOptions = [];
-
-  for (const {title, imageURL} of globalBadges) {
-    if (seenTitles.has(title)) {
-      continue;
-    }
-
-    seenTitles.add(title);
-    const keyword = badgeKeyword(title);
-    // label is required; mantine strips custom option props (title, imageURL) without it
-    badgeOptions.push({value: keyword, label: keyword, title, imageURL});
-  }
-
-  return badgeOptions.sort((a, b) => a.title.localeCompare(b.title));
-}
-
-let cachedBadgeOptions = null;
-let badgeOptionsPromise = null;
-
-async function fetchBadgeOptions() {
-  let globalBadges;
-  try {
-    globalBadges = await twitch.getGlobalBadges();
-  } catch (e) {
-    debug.log('failed to fetch twitch global badges', e);
-    badgeOptionsPromise = null;
-    return [];
-  }
-
-  cachedBadgeOptions = buildBadgeOptions(globalBadges);
-  return cachedBadgeOptions;
-}
-
-function useBadgeOptions(enabled) {
-  const [badgeOptions, setBadgeOptions] = useState([]);
-
-  useEffect(() => {
-    if (!enabled || cachedBadgeOptions != null) {
-      return undefined;
-    }
-
-    if (badgeOptionsPromise == null) {
-      badgeOptionsPromise = fetchBadgeOptions();
-    }
-
-    let mounted = true;
-    badgeOptionsPromise.then((options) => {
-      if (!mounted) {
-        return;
-      }
-
-      setBadgeOptions(options);
-    });
-
-    return () => {
-      mounted = false;
-    };
-  }, [enabled]);
-
-  return cachedBadgeOptions ?? badgeOptions;
 }
 
 function renderBadgeOption({option}) {
